--- app/src/main/java/com/techcourse/RenderingException.java
@@ -1,8 +1,0 @@
-package com.techcourse;
-
-public class RenderingException extends RuntimeException {
-
-    public RenderingException(Throwable cause) {
-        super(cause);
-    }
-}
