+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/View.java
@@ -6,7 +6,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
 
 public interface View {
-    void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
 
-    String getName();
+    void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
 }
