+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JspForwardException.java
@@ -1,0 +1,8 @@
+package webmvc.org.springframework.web.servlet.view;
+
+public class JspForwardException extends RuntimeException {
+
+    public JspForwardException(Throwable cause) {
+        super(cause);
+    }
+}
