+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JspView.java
@@ -1,11 +1,13 @@
 package webmvc.org.springframework.web.servlet.view;
 
+import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.View;
 
+import java.io.IOException;
 import java.util.Map;
 
 public class JspView implements View {
@@ -21,19 +23,29 @@ public class JspView implements View {
     }
 
     @Override
-    public void render(final Map<String, ?> model, final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        // todo
+    public void render(final Map<String, ?> model, final HttpServletRequest request,
+                       final HttpServletResponse response) throws Exception {
+        if (viewName.startsWith(JspView.REDIRECT_PREFIX)) {
+            response.sendRedirect(viewName.substring(JspView.REDIRECT_PREFIX.length()));
+            return;
+        }
+        setupAttribute(model, request);
+        forward(request, response);
+    }
 
+    private void setupAttribute(Map<String, ?> model, HttpServletRequest request) {
         model.keySet().forEach(key -> {
             log.debug("attribute name : {}, value : {}", key, model.get(key));
             request.setAttribute(key, model.get(key));
         });
-
-        // todo
     }
 
-    @Override
-    public String getName() {
-        return this.viewName;
+    private void forward(HttpServletRequest request, HttpServletResponse response) {
+        try {
+            final var requestDispatcher = request.getRequestDispatcher(viewName);
+            requestDispatcher.forward(request, response);
+        } catch (ServletException | IOException e) {
+            throw new JspForwardException(e);
+        }
     }
 }
