+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/RequestMappingHandlerAdapterTest.java
@@ -67,7 +67,8 @@ class RequestMappingHandlerAdapterTest {
 
         // then
         ModelAndView modelAndView = requestMappingHandlerAdapter.handle(request, response, handlerExecution);
-        assertThat(modelAndView.getView().getName()).isEqualTo("viewName");
+
+        assertThat(modelAndView).isNotNull();
     }
 
     @RequestMapping("/prefix")
